+++ src/main/java/seedu/ibook/model/item/UniqueItemList.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.ibook.model.product.exceptions.ItemNotFoundException;
+import seedu.ibook.model.item.exceptions.ItemNotFoundException;
 
 /**
  * A list of items that enforces uniqueness between its elements and does not allow nulls.
